# nitro2/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Chatroom:
    properties:
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      created_userid:
        type: string
        x-go-name: CreatorUserID
      display_name:
        description: the chatroom display_name
        type: string
        x-go-name: DisplayName
      enabled:
        type: boolean
        x-go-name: Enabled
      header:
        type: string
        x-go-name: Header
      id:
        description: the id for the chatroom
        minimum: 1
        type: string
        x-go-name: ID
      name:
        description: the name for the chatroom
        type: string
        x-go-name: Name
      purpose:
        type: string
        x-go-name: Purpose
      type:
        description: the chatroom type
        type: string
        x-go-name: Type
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    required:
    - name
    - display_name
    - type
    title: ChatroomRequest represents a muc chatroom response.
    type: object
    x-go-package: github.com/shizhongwang/myswagger/nitro2/data
  ChatroomMember:
    properties:
      chatroom_id:
        description: the id for the chatroom
        minimum: 1
        type: string
        x-go-name: ChatroomID
      lastviewed_at:
        format: date-time
        type: string
        x-go-name: LastViewedAt
      role:
        description: |-
          the user role in the chatroom

          required: failse
        type: string
        x-go-name: Role
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
      user_id:
        description: the id for the user
        type: string
        x-go-name: UserID
    required:
    - user_id
    title: ChatroomMemberRequest represents a muc chatroom response.
    type: object
    x-go-package: github.com/shizhongwang/myswagger/nitro2/data
  ChatroomMemberRequest:
    properties:
      role:
        description: |-
          the user role in the chatroom

          required: failse
        type: string
        x-go-name: Role
      user_id:
        description: the id for the user
        type: string
        x-go-name: UserID
    required:
    - user_id
    title: ChatroomMemberRequest represents a muc chatroom request.
    type: object
    x-go-package: github.com/shizhongwang/myswagger/nitro2/data
  ChatroomMembers:
    description: ChatroomMembers defines a slice of ChatroomMemberRequest
    items:
      $ref: '#/definitions/ChatroomMember'
    type: array
    x-go-package: github.com/shizhongwang/myswagger/nitro2/data
  ChatroomRequest:
    properties:
      display_name:
        description: the chatroom display_name
        type: string
        x-go-name: DisplayName
      enabled:
        type: boolean
        x-go-name: Enabled
      header:
        type: string
        x-go-name: Header
      name:
        description: the name for the chatroom
        type: string
        x-go-name: Name
      purpose:
        type: string
        x-go-name: Purpose
      type:
        description: the chatroom type
        type: string
        x-go-name: Type
    required:
    - name
    - display_name
    - type
    title: ChatroomRequest represents a muc chatroom request.
    type: object
    x-go-package: github.com/shizhongwang/myswagger/nitro2/data
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/shizhongwang/myswagger/nitro2/handlers
  ValidationError:
    description: ValidationError is a collection of validation error messages
    properties:
      messages:
        items:
          type: string
        type: array
        x-go-name: Messages
    type: object
    x-go-package: github.com/shizhongwang/myswagger/nitro2/handlers
info:
  description: Documentation for ChatroomRequest API
  title: of ChatroomRequest API
  version: 1.0.0
paths:
  /chatroomMembers:
    get:
      description: Return a list of ChatroomMembers from the database
      operationId: listChatroomMembersAll
      responses:
        "200":
          $ref: '#/responses/chatroomMembersResponse'
      tags:
      - chatroomMembers
    post:
      description: Create a new chatroomMember
      operationId: createChatroomMembers
      parameters:
      - description: |-
          ChatroomMembersRequest data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/ChatroomMemberRequest'
      responses:
        "200":
          $ref: '#/responses/chatroomMemberResponse'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - chatroomMembers
  /chatroomMembers/{chatroomid}:
    get:
      description: Return a list of ChatroomMembers from the database
      operationId: listChatroomMembersByChatroom
      parameters:
      - description: The id of the chatroomMember for which the operation relates
        in: path
        name: chatroomid
        required: true
        type: string
        x-go-name: ChatroomID
      responses:
        "200":
          $ref: '#/responses/chatroomMembersResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - chatroomMembers
  /chatroomMembers/{userid}:
    get:
      description: Return a list of ChatroomMembers from the database
      operationId: listChatroomMembersByUserID
      parameters:
      - description: By using the userid to search for the list of chatroomMembers,
        in: path
        name: UserID
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/chatroomMembersResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - chatroomMembers
  /chatrooms:
    get:
      description: Return a list of Chatrooms from the database
      operationId: listChatrooms
      parameters:
      - description: |-
          Currency used when returning the price of the chatroom,
          when not specified currency is returned in GBP.
        in: query
        name: Currency
        type: string
      responses:
        "200":
          $ref: '#/responses/chatroomsResponse'
      tags:
      - chatrooms
    post:
      description: Create a new chatroom
      operationId: createChatroom
      parameters:
      - description: |-
          ChatroomRequest data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/ChatroomRequest'
      responses:
        "200":
          $ref: '#/responses/chatroomResponse'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - chatrooms
    put:
      description: Update a chatrooms details
      operationId: updateChatroom
      parameters:
      - description: Newly created chatroom
        in: body
        name: Body
        schema:
          $ref: '#/definitions/Chatroom'
      responses:
        "200":
          $ref: '#/responses/chatroomResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
      tags:
      - chatrooms
  /chatrooms/{id}:
    delete:
      description: Update a chatrooms details
      operationId: deleteChatroom
      parameters:
      - description: The id of the chatroom for which the operation relates
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - chatrooms
    get:
      description: Return a list of Chatrooms from the database
      operationId: listSingleChatroom
      parameters:
      - description: |-
          Currency used when returning the price of the chatroom,
          when not specified currency is returned in GBP.
        in: query
        name: Currency
        type: string
      - description: The id of the chatroom for which the operation relates
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/chatroomResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - chatrooms
produces:
- application/json
responses:
  chatroomMemberResponse:
    description: Data structure representing a single chatroomMember
    schema:
      $ref: '#/definitions/ChatroomMember'
  chatroomMembersResponse:
    description: A list of chatroomMembers
    schema:
      $ref: '#/definitions/ChatroomMembers'
  chatroomResponse:
    description: Data structure representing a single chatroom
    schema:
      $ref: '#/definitions/Chatroom'
  chatroomsResponse:
    description: A list of chatrooms
    schema:
      items:
        $ref: '#/definitions/ChatroomRequest'
      type: array
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  errorValidation:
    description: Validation errors defined as an array of strings
    schema:
      $ref: '#/definitions/ValidationError'
  noContentResponse:
    description: No content is returned by this API endpoint
schemes:
- http
swagger: "2.0"